validate.py: handle missing keys gracefully
As mentioned in bsc#1000620, if the pillar data is largely
nonexistent (due to a misconfiguration in that bug),
validate.py will bail out with KeyError exceptions.  Because
it throws an exception rather than returning False in this
case, the "validate failed" state in ceph/stage/deploy.sls
doesn't trigger, and salt attempts to run all the other
deploy states which fail rather spectacularly at that point.

This commit makes validate.py cope properly with missing
keys, although some of the error text when the keys aren't
there could still be improved (empty strings get interpolated
in a couple of places).

Signed-off-by: Tim Serong <tserong@suse.com>