Makefile: create install target
This moves all the `install` invocations from the specfile to
the Makefile, and collapses as many as possible with wildcards.
Where wildcards are not used, it's usually because there's
still a custom.sls file in the source tree that we don't want
to be installed.

Having `make install` means it's easy to try DeepSea on
arbitrary non-RPM-based distros (`git clone ; make install`)
and should also hopefully help with packaging for other distros.

Signed-off-by: Tim Serong <tserong@suse.com>