wf-auth: element to authenticate with salt-api backend
Authentication is not in the traditional sense.  To speak to the salt-api
backend, one requires the salt-api node, port, userid, password and the
external authentication mechanism employed.  We pull this information from
the user via an auth dialog and issue a salt-api call (via the <salt-api>
element).  On success, we store this data in sessionStorage.wfauth to be
utilized later by other elements to communicate with the salt-api
backend (ie. utility elements that do not yet exist).

Failing to authenticate will clear sessionStorage.wfauth while retaining
the auth dialog entries for easier correction by the user.  Clicking the
logout button will also clear out all sessionStorage.wfauth data, along
with the auth dialog.

On a successful 'login', reloading the page will reinstantiate credentials
from sessionStorage.wfauth.

Note: We may wish to replace the 'logout' button with a 'clear' button
that is always enabled.  This would allow the user to clear out the auth
dialog input even when authentication failed.  A minor cosmetic thing.

Signed-off-by: Karol Mroz <kmroz@suse.de>