salt-api/wf-salt-api: add wraper around generic salt-api
To be used by elements that rely on salt-api for everything other than
authentication.  Fewer attributes needed, and specific handling of 401
events (ie. in the case the admin comes back to an active view).
Yes, there is a bit of duplication between this and wf-auth with respect to
scraping sessionStorage.wfauth.

Signed-off-by: Karol Mroz <kmroz@suse.de>