wolffish-app: remove auth state checks when routing pages
I didn't like this solution to begin with.  Now we simply allow failed
wf-salt-api ajax requests with status 0 (server down) or 401 (unauthorized)
to fire a 'wf-salt-api-error-401' event that will open the wf-auth dialog
in modal mode and prompt the user to login.  It also allows, after the
first login, for a user to navigate the application even if the backend
salt-api server has gone down.  I think it's a better UX.  We'll see.

Fixes: #11

Signed-off-by: Karol Mroz <kmroz@suse.de>