Add minimal Debian build files
This should be the minimum set of files necessary to build for
Debian and/or Ubuntu.  To use this on build.opensuse.org, the
changelog, compat, control and rules files need to be included
in the OBS project, renamed to debian.changelog, debian.compat,
debian.control and debian.rules.  OBS also requires a deepsea.dsc
file, which AIUI would ordinarily be generated by dpkg-source,
which is why I haven't included it in this commit.  For an
example build on OBS, see:

  https://build.opensuse.org/package/show/home:tserong:branches:home:swiftgist/deepsea

Note that this will build for Debian 8.0 and Ubuntu 16.x, but
not for Debian 7.0 or Ubuntu 14.x, as the latter two don't seem
to have salt-master available.

Also note that this is missing deepsea.postinst, which we
probably want to add at some point to run the same commands
as are in the %post section of the RPM spec file (for some
reason I couldn't get the postinst script to be included
when doing test builds on OBS, so I thought it best to leave
that for later).

Signed-off-by: Tim Serong <tserong@suse.com>