Copy ntp-formula in entirety
We are ingesting SaltStack's premade NTP formula for the time service
feature to leverage the prior art. We are not installing this and using
it as a formula traditionally would be because we do not want to
interfere with the same (or a similar) NTP formula a user may already
have installed.

This is a first, full-directory (minus git metadata) copy of SaltStack's
ntp-formula before we make modifications.

Copied from https://github.com/saltstack-formulas/ntp-formula
at commit c0e05121abf1ca4b665699fe610da7c32ad6e2b7.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>