osd.py: run partprobe on the newly created partition
If only run on the device file (without partition number) there is a
chance that partprobe returns with success, but dd after does not find
the partition (his happens when two partitions are created in a row on
the same disk device - dd fails to find the second created partition).
If partprobe is run on the created partition everything works.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>