cephinspector: skip "disks" with no partitions
When ceph-disk encounters /dev/sr0, that's included in its output, but the
dict for that device has no "partitions" key, which results in a traceback
(KeyError: 'partitions') a couple of lines later.  That traceback then lands
in the generated YAML file (yuck).

In general, we need a good way of dealing with failures like this from
cephinspector so that we don't accidentally do Stupid Things[TM] later :-/

Signed-off-by: Tim Serong <tserong@suse.com>