build: Add key cache directories to package
When using the populate runner on the master to import an existing
cluster, the salt user needs to be able to write keys to the various
cache directories.  For a regular deployment (i.e. not an import),
these directories are created at runtime anyway with mode 0600,
user and group salt, but that doesn't happen until stage 2, which
is too late for the import case.  This change adds these cache
directories to the package with appropriate ownership and perms.

Signed-off-by: Tim Serong <tserong@suse.com>