stage 1: network discovery revisited
This new discovery algorithm uses two new heuristics to find the
main public network and the cluster network.

The heuristic to find the public networks works like this:
If a network matches one of the "hostname -i" of the minions, then
the network is considered as a public network.

The heuristic to find the cluster networks works like this:
If a network that has more than one minion does not match any
of the salt-master interface addresses, then the network is
considered as a cluster network.

Otherwise, the discovery fails to the already existing behavior.

Signed-off-by: Ricardo Dias <rdias@suse.com>