qa: add ceph-test suite with librbd test
The "ceph-test" suite is for tests that require the ceph-test package to be
installed on all the test nodes.

The librbd.sh test in that suite deploys a cluster in which ceph_test_librbd
can be run on the client node. It does not actually execute ceph_test_librbd,
though, because that binary generates too much output for cmd.run.

Signed-off-by: Nathan Cutler <ncutler@suse.com>