qa: validate: echo validation report to log
Don't know exactly why (presumably the runner is a subprocess and its stdout is
not captured), but the validation report does not appear in the teuthology log.

Not being able to see the validation messages makes it hard to debug test
failures. Therefore, echo these messages to the logger at the appropriate
severity level.

Fixes: https://github.com/SUSE/DeepSea/issues/611
Signed-off-by: Nathan Cutler <ncutler@suse.com>