cephprocesses: enhance check() function
Replace status.pid with psutil.process_iter().  Allows for proper
matching of process names.

Add a results flag to track which processes are up/down.  This will be
useful when wanting to determine if, for example, _all_ Ceph processes
meant to be running on a given node are down.

ses5-cephsubvol-2.lan:
    ----------
    down:
        - radosgw
    up:
        ----------
        ceph-mgr:
            - 17456
        ceph-mon:
            - 17455
        ceph-osd:
            - 17468

By default, check() will still return False if a Ceph process meant to be
running on a given node is down, and True otherwise.

Add a quiet flag to not log an error when a process is isn't running.
This is useful when checking if all processes are down (where a dead
process is not an error).

Also, fix the 'openattic' process name.

Signed-off-by: Karol Mroz <kmroz@suse.de>