populate: support multiple public, cluster networks in engulf (bsc#1076010)
This removes the assumption in the engulfer that there can be no more
than one public or cluster network.  Without this change, running engulf
on a cluster with (e.g.) public_network = 10.27.21.0/24, 10.27.20.0/24
will result in the following error:

  [ERROR   ] Failed to determine cluster's public_network.
  [ERROR   ] Failed to replace derived public_network with public_network
             of existing cluster.

Note that this still leaves the public_network and cluster_network settings
as strings in the pillar data (changing the pillar data to be an array would
have been somewhat more invasive).

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1076010
Fixes: https://github.com/SUSE/DeepSea/issues/582
Signed-off-by: Tim Serong <tserong@suse.com>