Invoke `find` to fix salt job cache permissions
When there's a lot of files in the salt job cache, using salt's
builtin file.directory construct is remarkably slow.  Let's start
with about 10,000 files in the cache, all correctly owned:

  # find /var/cache/salt/master/jobs | wc -l
  9920
  # find /var/cache/salt/master/jobs -user root | wc -l
  0

Then, run ceph.stage.2:

  # salt-run state.orch ceph.stage.2
  [...]
  Summary for admin.ceph_master
  -------------
  Succeeded: 14 (changed=10)
  Failed:     0
  -------------
  Total states run:     14
  Total run time:  409.141 s

That's 409 seconds, or a bit under seven minutes.  Now, with this
change applied, which instead invokes `find ... -exec chown ...`,
it only takes a minute twenty, which is quite a bit better:

  # salt-run state.orch ceph.stage.2
  [...]
  Summary for admin.ceph_master
  -------------
  Succeeded: 14 (changed=10)
  Failed:     0
  -------------
  Total states run:     14
  Total run time:   82.443 s

Fixes: https://github.com/SUSE/DeepSea/issues/1131
Signed-off-by: Tim Serong <tserong@suse.com>