qa: fix ceph_version_test
This function could have produced a false positive if both the RPM and Ceph version strings
failed the regex match. (In that case, both the variables would have been empty and the
equality test might have passed.)

With this commit, the test explicity checks that the regex match actually produced a string
that can be usefully compared.

Signed-off-by: Nathan Cutler <ncutler@suse.com>