build/ops: streamline qa %files section
Merely mentioning the top-level directory is sufficient to bring in that
directory and everything under it.

This is a Big Stick, and not to be used lightly, but since (a) the deepsea-qa
package is not intended for wide consumption and (b) the directories in
question contain only QA-related stuff, it seems OK to use here.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit b3915d685268e792dd26b409012ef68bbea06dbd)

Conflicts:
	deepsea.spec.in (trivial resolution)