qa/deploy: handle possibility that deepsea-cli is not installed
When installing DeepSea from packages, the "deepsea-cli" package might
not get installed (it is not strictly required), and depending on which tests
are being run, the absence of the CLI might not be a problem.

In any case, there is no reason why failure of "deepsea --version" should
fail the entire test.

Signed-off-by: Nathan Cutler <ncutler@suse.com>