tests: disable CLI unit tests in Jenkins PR CI
The CLI unit tests use docker, which does not work in Jenkins PR CI, so ignore
these tests when running via Jenkins.

This is not cherry-picked from master because in master the change was made
as part of a more complicated patchset involving py-coverage and codecov, which
is not designated for backport.

Signed-off-by: Nathan Cutler <ncutler@suse.com>