qa: reinstate health-openattic.sh
The Jenkins CI expects to be able to deploy a Ceph cluster with openATTIC by
running a script called health-openattic.sh in the qa/suites/basic directory of
the DeepSea source tree.

In the recent QA refactor for SES5 (https://github.com/SUSE/DeepSea/pull/1289)
this script got moved to the top directory by mistake.

This commit drops the old versions of the health-igw.sh and health-openattic.sh
scripts and reimplements health-openattic.sh as a wrapper around health-ok.sh.

NOTE: This cannot be cherry-picked from master because openATTIC has been
replaced by the Ceph MGR Dashboard in the meantime.

Signed-off-by: Nathan Cutler <ncutler@suse.com>