qa/deploy: use jq instead of python in salt_api_test
Also, use "tee" and a temporary file to get the output from curl
displayed in the log. (As long as the curl command produces valid JSON,
it's enough to pipe the output into jq, because jq will display it.
If curl produces anything else, however, jq will *not* display what
it was trying to parse.)

Also, since the curl output does not have a trailing newline, we need to
provide one ourselves, lest we get this in the log:

{"clients": ["local", "local_async", "local_batch", "local_subset", "runner", "runner_async", "ssh", "wheel", "wheel_async"], "return": "Welcome"}Salt API test: END

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit a5a823074d7590366f4899d7991c72a4cdb1fd87)