qa: run 'remove' functest last
The remove functest removes the first two OSDs, but doesn't wipe the
disks.  That fine, but it means if we run the engulf functest *after*
remove, it will detect those still-existing disks, and generate a storage
profile which no longer matches the default profile, and thus cause the
test to fail.

Signed-off-by: Tim Serong <tserong@suse.com>