Makefile: Check for correct ID in /etc/os-release
Without this patch, is not possible to install DeepSea using openSuse
Leap 15.0, since the /etc/os-release ID field is "opensuse-leap" [1],
so the Makefile was checking for "opensuse", "opensuse-tumbleweed" and
"sles". Now zypper is able to install sal-api and other necessary
packages since Leap is detected and the PKG_INSTALL variable is
populated.

[1]: https://en.opensuse.org/SDB:Find_openSUSE_version

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>