Add mgr_orch runner to support ceph-mgr deepsea orchestrator module
This runner is intended to provide a 1:1 mapping between the ceph-mgr
orchestrator API, and DeepSea internals.  This enables us to keep the
ceph-mgr orhcestrator module quite simple, and do any necssary data
aggregation here inside DeepSea itself.  This commit introduces only
two methods so far (get_inventory and describe_service) for querying
nodes, devices and services.

I've also added a "blank" field in cephdisks.py to indicate whether
a disk is empty or not (the orchestrator API wants to know this).

Signed-off-by: Tim Serong <tserong@suse.com>