Makefile: pass --long option to git describe
Before this commit, we were generating the version number using
plain "git describe", which (according to the man page)

    "finds the most recent tag that is reachable from a commit. If the tag
    points to the commit, then only the tag is shown."

This had the undesirable effect of sometimes producing version numbers
that look like this:

    0.8.9+2.g14f2f94e

and other times like this:

    0.8.9

Depending on whether HEAD pointed to a tag or not. By passing the --long
option, we assure that the long format is used always.

Signed-off-by: Nathan Cutler <ncutler@suse.com>