metapackage: add test.nop stanza
Without this, DeepSea fails in openSUSE because the osfullname grain contains
the string "Leap". As a result, the Jinja conditional evaluates to false and
the file becomes empty, causing Salt to generate a bizarre error.

(This is not cherry-picked from master because the fix for this issue in master
is too complicated.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>