Remove RGW demo bucket creation
With the support of IPv6, the parsing in rgw.py would need to be much
more sophisticated to handle civetweb, beast and the various keyword
parameters for either.

After discussion with Abhi and Ricardo D., the demo buckets were only
necessary for the standalone openATTIC and ganesha.  With oA as the Ceph
dashboard and a newer implementation for configuring ganesha from the
Ceph dashboard, these buckets are not needed.

Signed-off-by: Eric Jackson <swiftgist@gmail.com>