conf: rgw: drop rgw dns name configurable
This is only useful for s3 clients which don't support path style access, and
even in those cases there needs to be an additional dnsmasq setting or the like
that can correctly redirect `<bucketname>.<rgw-endpoint>` to `<rgw-endpoint>`.
For the remainder of use cases if this is configured incorrectly it always leads
to bugs for all s3 clients and it is hard to get this right as many rgws may be
terminated with a load balancer in which case we need the endpoint to be that.
So moving forward we'd make this option manually configurable for those sites
that really need the vhost style access as this requires some knowledge on the
s3 clients and their site dns setting which is hard to predict correctly anyway

Fixes: #842
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>