utils/status: handle down nodes gracefully
When querying grains and running commands, any nodes that are down will
result in a dict value of False, and a 15 second delay (which is horrible).
This commit detects down nodes on the first grains.get, then subsequent
queries exclude those nodes, and we return "Unknown (node down?)" instead
of False.

Signed-off-by: Tim Serong <tserong@suse.com>