cephfs: fix pool creation prerequisite
With unless the state is executed when one condition returns false. For
example it'll try to create the cephfs_data pool if the pool is present
but the cephfs instance is not yet.
Negating the condition and using onlyif yields the expected behaviour.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>
(cherry picked from commit 8667b12adbd66a773e29cc7c59e97e496a586b05)