runners/upgrade: Add generate_service_specs function
This adds YAML service spec export suitable for applying to
the ceph orchestrator, based on DeepSea configuration.  Call
via `salt-run upgrade.generate_service_specs > specs.yaml`.
This will generate service specs for mon, mgr, osd, crash,
node-exporter, prometheus, grafana and alertmanager.  It does
not include mds or gateways, as these should be redeployed,
and in any case may require reconfiguration or extra
parameters which we may not be able to guess at, so it's
safer to leave these services up to the user to redeploy.

Signed-off-by: Tim Serong <tserong@suse.com>