Split OSD restart and actiavtion check loop in two
By restarting all OSDs, then going back and checking if they're
all active after all the restarts are complete, we should shave
a couple of seconds off most of the checks (this becomes quite
noticeable once you have more than a few nodes with more than
a handful of OSDs each).  This commit also moves the sleep in
osd.wait_until_available() to the end of the loop, to give a
further speedup in the happy case.

Signed-off-by: Tim Serong <tserong@suse.com>