Add osd_restart_delay option
The previous two commits address the problem seen in bsc#1185422,
i.e. given the addition of osd.wait_until_available(), we should now
never proceed to restart the OSDs on the next node, until all the
OSDs on the current node are up and in.  In case there is still a
concern that the whole restart process may be too fast, this commit
adds the ability to inject a delay after restarting OSDs on each
node.  By default, the delay is zero, so has no effect.  To activate
this feature, set "osd_restart_delay: 60" (or however many seconds
you would like to delay) to /srv/pillar/ceph/stack/global.yml

Signed-off-by: Tim Serong <tserong@suse.com>