Fixes no UCST response after 1 seconds
Although this code seems to have been working for about 5 years, it
doesn't so anymore for me. I'm not sure what happened, but a second is
insufficient now, and that reproducibly.

When changing the time to wait for a response up to two seconds, it
works.

When the `-t <int>` argument is removed, a default of 5 seconds is used,
so I decided to go with that.

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>