osd: handle `ceph osd ok-to-stop` output from ceph v14.2.22
Prior to https://github.com/ceph/ceph/commit/a45bfebb, the
`ceph osd ok-to-stop` command output a long, human readable
string.  DeepSea was looking at this string to see if OSDs
really were OK to stop.  Since that commit, the output of
this command is now a chunk of JSON, which obviously can't
work with that string comparison, so let's try to parse the
JSON first, in case we're using the latest Nautilus release,
and if that fails, fall back to the previous check in case
we're still running against an older Ceph.

Signed-off-by: Tim Serong <tserong@suse.com>