Bump OSD available timeout to 10mins, make it configurable (bsc#1192520)
When restarting OSDs, we wait for them to become available before
proceeding to the next OSD.  Previosly, this wait was in a loop with
a timeout of 60 seconds.  This commit increases the default timeout
to 600 seconds (10 minutes), and also makes it configurable in case
it's necessary to change for any site-specific reason.  If you need
a different timeout, specify it in /srv/pillar/ceph/stack/global.yml,
for example:

osd_available_timeout: 300

Note that the way the loop is implemented, it will return immediately
if the OSD is already active.  If it's not active, it will retry
the check periodically until the timeout is exceeded.  In order to
return quickly for OSDs that start quickly, the first sleep is 1
second, then 2 seconds, then 4, 8, 16, 32, but after that it reverts
to sleeping 60 seconds per iteration.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1192520
Signed-off-by: Tim Serong <tserong@suse.com>