Use python3.6 for tox and travis
Commit f8a62692 switched to using generic python3, allegedly because
Tumbleweed had python 3.7, not 3.6.  I assume that was true at the
time, because I wrote it, but given that python 3.6 *is* currently
available on Tumbleweed, perhaps I was mistaken.  Anyway, it makes
the most sense to force testing with python3.6, because that's what
we deploy against with SES6 on SLE 15 SP1.

This commit also:

- Forces the use of isort < 5 when running lint, because version 5
  produces a pile of "module 'isort' has no attribute 'SortImports'"
  errors, and no actual useful lint output.
- Gets rid of the file permission checks when running in Travis CI,
  because Travis sets umask 002, which messes up the file permissions
  during `git clone`, and causes all the tests to fail.

Signed-off-by: Tim Serong <tserong@suse.com>