Ensure upgrade.status lists nodes with non-ceph roles (bsc#1195366)
When sorting nodes for display, the upgrade.status runner was only
taking into account the nodes with roles 'master', 'mon', 'mgr',
'storage', 'mds', 'rgw', 'igw' and 'ganesha'.  It wasn't handling
the 'prometheus' or 'grafana' roles.  If the latter two roles are
assigned to nodes that also have one of the "main" roles, this is
not a problem, but if prometheus and/or grafana are deployed on
separate nodes by themselves (with no other roles), those nodes
don't appear in the list of nodes to upgrade, leading one to think
that everything is upgraded, when it's potentially not.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1195366
Signed-off-by: Tim Serong <tserong@suse.com>