Extra information when PGs take too long to be in active+clean state
When running rebuild.node it waits for PGs to be in the active+clean
state.
If PGs are, for example scrubbing, the user only gets the message: "Waiting for PGs to recover..."

This branch adds extra information, showing the actual state of PGs and,
in the case that they're scrubbing, it sugests to disable scrubbing and
trying.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1194807
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>