Don't use punctuation when generating dashboard admin password
If the string generated by random.get_str includes punctuation, it can
break the parsing of the jinja template and/or screw up the line that
echoes the password into `ceph dashboard ac-user-create`.  Salt 3004
added the ability to limit what characters are used in the password,
so let's turn off punctuation to avoid this problem.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1200794
Signed-off-by: Tim Serong <tserong@suse.com>